+++ remove.sh
@@ -13,15 +13,15 @@ banner() {
     printf "\033[36m    |  | |__] |  | |\ |  |  |  |    |\/| |  | |  \ \033[0m\n"
     printf "\033[32m    |__| |__] |__| | \|  |  |__|    |  | |__| |__/ \033[0m\n"
     printf "\033[0m\n"
-    printf "     \033[32mA modded gui version of ubuntu for Termux\033[0m\n"
+    printf "     \033[32mA modded gui version of Debian for Termux\033[0m\n"
     printf "\033[0m\n"
 
 }
 
 package() {
     echo -e "${R} [${W}-${R}]${C} Purging packages..."${W}
-    proot-distro remove ubuntu && proot-distro clear-cache
-    rm -rf $PREFIX/bin/ubuntu
+    proot-distro remove debian && proot-distro clear-cache
+    rm -rf $PREFIX/bin/debian
     sed -i '/pulseaudio --start --exit-idle-time=-1/d' ~/.sound
     sed -i '/pacmd load-module module-native-protocol-tcp auth-ip-acl=127.0.0.1 auth-anonymous=1/d' ~/.sound
     echo -e "${R} [${W}-${R}]${C} Purging Completed !"${W}